Make getMissing available in lib
