+++ package.json
@@ -60,6 +60,7 @@
     "fs-extra": "11.3.5",
     "highlight.js": "^11.11.1",
     "json-stable-stringify": "1.3.0",
+    "jsonc-parser": "^3.3.1",
     "slugify": "1.6.9",
     "tsx": "^4.22.4",
     "ua-parser-js": "2.0.10",
@@ -105,7 +106,6 @@
     "foreman": "^3.0.1",
     "globals": "^17.6.0",
     "json3": "3.3.3",
-    "jsonc-parser": "^3.3.1",
     "listr2": "10.2.1",
     "minimatch": "10.2.5",
     "mocha": "11.7.6",
